Split tutorial-navigator file, updated build steps and sample
