+++ bower.json
@@ -18,6 +18,7 @@
     "tests"
   ],
   "dependencies": {
-    "auth0-styleguide": "git://github.com/auth0/styleguide.git#2.1.2"
+    "auth0-styleguide": "git://github.com/auth0/styleguide.git#2.1.2",
+    "jquery": "https://github.com/jquery/jquery.git#1.9"
   }
 }
