+++ lib/index.jsx
@@ -1,0 +1,10 @@
+/** @jsx React.DOM */
+var Navigator = require('./tutorial-navigator')
+
+TutorialNavigator = (function() {
+  return {
+    init: Navigator
+  }
+})();
+
+module.exports = TutorialNavigator;
